LayoutTests/editing/style/fore-color-by-name-expected.txt


"world" should be red:
| "hello <#selection-anchor>"
| <font>
|   color="#ff0000"
|   "world<#selection-focus>"

there should be no span or font elements:
| "<#selection-anchor>hello "
| "world<#selection-focus>"
